third_party/blink/web_tests/flag-specific/highdpi/fonts/data-url-sync-loading-expected.txt

This is a testharness.js-based test.
FAIL Inline web fonts should be loaded synchronously when used. assert_equals: expected 250 but got 253
Harness: the test ran to completion.
